MIBE-58: feat: add relevant attributes per selectable tag
